Team health indicator Fixes #57
