+++ package.json
@@ -14,7 +14,6 @@
     "firebase": "^8.2.6",
     "lodash": "^4.17.20",
     "react": "^16.14.0",
-    "react-chord-diagram": "^1.7.0",
     "react-dom": "^16.14.0",
     "react-equation": "^0.2.1",
     "react-firebaseui": "^4.1.0",
