+++ src/Authors.js
@@ -127,11 +127,15 @@ export default function Authors() {
                 <ResponsiveContainer width='100%' aspect={3 / 1} margin={{ top: 0, right: 0, left: 0, bottom: 0 }}>
                     <AreaChart
                         data={
-                            author.Ratings.map(r => ({
-                                CreatedAt: new Date(r.CreatedAt).getTime(),
-                                Value: r.Value,
-                                WorkId: r.WorkId
-                            }))
+                            author.Ratings.map(
+                                r => (
+                                    {
+                                        CreatedAt: new Date(r.CreatedAt).getTime(),
+                                        Value: r.Value,
+                                        WorkId: r.WorkId
+                                    }
+                                )
+                            )
                         }
                         margin={{ top: 0, right: 0, left: 0, bottom: 0 }}>
                         <Tooltip wrapperStyle={{ backgroundColor: '#fff', border: '1px solid #dcdcdc' }} content={<CustomTooltip />} />
